2.9 Commit
Created the_gear.py again and brought everything over to it.
